Fix compiler warnings in BufferDisplay.cpp
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@145150 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 6fc3ca6572e247b12ac64dd8f9e80a37bd1f8829

+++ Online/Online/ROMon/src/BufferDisplay.cpp
@@ -106,7 +106,7 @@ void BufferDisplay::update(const void* data) {
           for (Clients::const_iterator ic=clients.begin(); ic!=clients.end(); ic=clients.next(ic))  {
             Clients::const_reference c = (*ic);
             cnam = (char*)c.name;
-	    int mx_len = 20;
+	    size_t mx_len = 20;
             if ( ::strlen(cnam)>mx_len && (p=::strchr(cnam,'_')) ) {// Cleanup name if too long to be displayed properly...
               cnam = ++p;
               if ( ::strlen(cnam)>mx_len && (p=::strchr(cnam,'_')) ) {
